+++ Vasilyev/5) [DinArray, DinArray2, Vector, VectorPlusMinusAvg] - 30.09.16/VectorPlusMinusAvg.cpp
